requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
pillow==12.1.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
python-dotenv==1.2.1
djangorestframework-simplejwt
drf-spectacular==0.29.0
